% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_stat.R
\name{d_stat}
\alias{d_stat}
\title{calculate the energy statistics from balanced designed samples}
\usage{
d_stat(size, group, index, distance)
}
\arguments{
\item{size}{the sample size}

\item{group}{the number of the groups of the samples}

\item{index}{for permutation test, indicate how to divide the samples into groups}

\item{distance}{the distance matrix}
}
\value{
energy statistics used for multivariate test for equal distribution
}
\description{
This fucntion calculate the statistics to test multrivariate test for equal distributions based on energy method
}
