% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_correlation.R
\name{dCov}
\alias{dCov}
\title{compute the distance covariance of two samples}
\usage{
dCov(x, y)
}
\arguments{
\item{x}{one of the sample ,can be a vector,a matrix}

\item{y}{the other sample,the dimension must agree with the x}
}
\description{
this function generate the statistic of the test for dependence based on distance
}
