% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_test.R
\name{equal_test}
\alias{equal_test}
\title{multivariate test for equal distribution}
\usage{
equal_test(dat, size, group, dim, R = 199)
}
\value{
p value from the permutation test
}
\description{
implement a permuation test, using energy method to calculate method for multivariate test for equal distribution
}
