% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_normal_test.R
\name{genirate_mix_normal}
\alias{genirate_mix_normal}
\title{genirate a population of mix normal distribution}
\usage{
genirate_mix_normal(n, ...)
}
\arguments{
\item{n}{number of iid rv}

\item{...}{mean, sd, pr of each normal population}
}
\description{
this function genirate n indepently identically distributed random variables
}
