% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_correlation.R
\name{ndCov2}
\alias{ndCov2}
\title{compute the distanc correlation of two samples which known as Rn}
\usage{
ndCov2(z, ix, dims)
}
\arguments{
\item{z}{the matrix of the two samples}

\item{ix}{the index of one of sample in the matrx}

\item{dims}{the dimension of the matrix z}
}
\description{
this function based on the dcov and give the final result
}
