% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_normal_test.R
\name{norm.plot}
\alias{norm.plot}
\title{plot the p.d.f of two normal distribution}
\usage{
norm.plot(mean1, sd1, mean2, sd2)
}
\arguments{
\item{mean1}{the mean of the first population}

\item{sd1}{the sd of the first population}

\item{mean2}{the mean of the second population}

\item{sd2}{the sd of the second population}
}
\description{
plot the p.d.f of two normal distribution
}
