% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_normal_test.R
\name{pow2p.a1.2H30}
\alias{pow2p.a1.2H30}
\title{estimation the power in a mixture probability
under alpha=0.1
m=300
n=30
two population}
\usage{
pow2p.a1.2H30(mean1, sd1, mean2, sd2)
}
\arguments{
\item{mean1}{the mean of the first population}

\item{sd1}{the sd of the first population}

\item{mean2}{the mean of the second population}

\item{sd2}{the sd of the second population}
}
\description{
estimation the power in a mixture probability
under alpha=0.1
m=300
n=30
two population
}
