% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_normal_test.R
\name{pow2p.plot}
\alias{pow2p.plot}
\title{plot the power function of a 2 population mixed nuomal distribution}
\usage{
pow2p.plot(A)
}
\arguments{
\item{A}{the power matric}
}
\description{
plot the power function of a 2 population mixed nuomal distribution
}
