% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_normal_test.R
\name{powt}
\alias{powt}
\title{estimation the power in a mixture probability}
\usage{
powt(alpha, m, n, ...)
}
\arguments{
\item{alpha}{the pointed value}

\item{m}{the number of replicates during Monte Carlo Methods for Estimation}

\item{n}{the size of the sample}

\item{...}{mean, sd, pr of each normal population}
}
\description{
estimation the power in a mixture probability
}
