% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_test.R
\name{equal_test}
\alias{equal_test}
\title{multivariate test for equal distribution}
\usage{
equal_test(dat, size, group, dim, R = 199)
}
\arguments{
\item{dat}{data should be a matrix with each row representing a point}

\item{size}{the sample size}

\item{group}{the number of the groups of the samples}

\item{dim}{the dimension of the point}
}
\value{
p value from the permutation test
}
\description{
implement a permuation test, using energy method to calculate method for multivariate test for equal distribution
}
